32bit lintian fix

+++ gcl/o/funlink.c
@@ -328,7 +328,7 @@ call_proc_new(object sym,ufixnum clp,ufixnum vld,void **link,ufixnum argd,object
     if (tp==t_function) {
       fprintf(stderr,"Warning: arg/val mismatch in call to %-.*s (%p) prevents fast linking:\n %ld %ld/%ld %d(%d)  %ld %d  %ld %d  %ld, recompile caller\n",
 	      (int)(type_of(sym)==t_symbol ? VLEN(sym->s.s_name) : 0),sym->s.s_name->st.st_self,sym,
-	      argd,fun->fun.fun_argd,
+	      argd,(long)fun->fun.fun_argd,
 	      vald,fun->fun.fun_neval,fun->fun.fun_vv,
 	      margs,fun->fun.fun_minarg,nargs,fun->fun.fun_maxarg,pushed);
       fflush(stderr);
